Create user model
That would be a lot easier to make it with devise gem, but it generates much
more code than I actually need for now. Then, for simplicity sake, I decided
to create a model and table on my own

Moreover, I ended up adding a gem installation by accident instead of creating
a commit for this addition. This gem will allow me to encrypt the password